+++ packages/react-ui/components/DateInput/helpers/InternalDateMediator.ts
@@ -53,9 +53,10 @@ export class InternalDateMediator {
   public inputKey(key: string, type: InternalDateComponentType | null, inputMode: boolean): InputKeyResult {
     const prevValue = this.iDate.get(type);
     if (type === null) {
-      type = this.getLeftmostType();
-      this.clear(type);
+      const leftMostType = this.getLeftmostType();
+      this.clear(leftMostType);
     }
+
     if (type !== InternalDateComponentType.Year) {
       this.iDate.cutOffExcess(type);
     } else {
@@ -84,11 +85,13 @@ export class InternalDateMediator {
   };
 
   public shiftDateComponent(type: InternalDateComponentType | null, step: number): boolean {
-    type = type === null ? this.getLeftmostType() : type;
+    const calculatedType = type ?? this.getLeftmostType();
+
     const iDate = this.iDate.clone();
-    const isValidRange = iDate.validate({ checks: [InternalDateValidateCheck.Range] });
     const start = iDate.getRangeStart();
     const end = iDate.getRangeEnd();
+
+    const isValidRange = iDate.validate({ checks: [InternalDateValidateCheck.Range] });
     if (!isValidRange) {
       // Удерживаем дату в заданном диапазоне
       if (start && InternalDateGetter.max([iDate, start]) === start) {
@@ -97,12 +100,13 @@ export class InternalDateMediator {
         iDate.duplicateOf(end);
       }
     } else {
-      const clone = iDate.clone().shift(type, step, { isRange: false, isLoop: true, isCutFeb: true });
+      const clone = iDate.clone().shift(calculatedType, step, { isRange: false, isLoop: true, isCutFeb: true });
       if (clone.validate({ checks: [InternalDateValidateCheck.Range] })) {
         iDate.duplicateOf(clone);
       }
     }
-    const changed = !this.iDate.isEqualComponentDate(type, iDate);
+
+    const changed = !this.iDate.isEqualComponentDate(calculatedType, iDate);
     this.iDate = iDate;
     return changed;
   }