+++ packages/react-ui/components/TooltipMenu/TooltipMenu.tsx
@@ -12,7 +12,7 @@ import { rootNode, TSetRootNode } from '../../lib/rootNode';
 
 export type TooltipMenuChildType = React.ReactElement<MenuItemProps | {} | MenuHeaderProps>;
 
-export interface TooltipMenuProps extends CommonProps {
+export interface TooltipMenuProps extends CommonProps, Pick<PopupMenuProps, 'onOpen' | 'onClose'> {
   children?: TooltipMenuChildType | TooltipMenuChildType[];
   /** Максимальная высота меню */
   menuMaxHeight?: number | string;
@@ -114,7 +114,9 @@ export class TooltipMenu extends React.Component<TooltipMenuProps> {
           header={this.props.header}
           footer={this.props.footer}
           positions={this.props.positions}
-          popupHasPin={true}
+          onOpen={this.props.onOpen}
+          onClose={this.props.onClose}
+          popupHasPin
           disableAnimations={this.props.disableAnimations}
         >
           {this.props.children}