+++ packages/react-ui/internal/Menu/Menu.styles.ts
@@ -5,9 +5,11 @@ export const styles = memoizeStyle({
   root(t: Theme) {
     return css`
       background: ${t.menuBgDefault};
+      border-radius: ${t.menuBorderRadius};
       box-sizing: content-box;
       overflow: auto;
-      padding: 0 0;
+      padding: 0 ${t.menuPaddingX};
+      border-radius: ${t.menuBorderRadius};
     `;
   },