+++ packages/react-ui/lib/events/keyboard/extractCode.ts
@@ -4,14 +4,16 @@ import { KeyboardEventCodes as Codes } from './KeyboardEventCodes';
 import { KeyboardKey, KeyboardMapKeys } from './KeyboardMapKeys';
 
 export function extractCode(e: React.KeyboardEvent<HTMLElement> | KeyboardEvent): Codes {
-  e = (e as React.KeyboardEvent<HTMLElement>).nativeEvent || e;
-  if (e.code) {
-    return e.code as Codes;
+  const event = (e as React.KeyboardEvent<HTMLElement>).nativeEvent || e;
+  if (event.code) {
+    return event.code as Codes;
   }
-  const keyCode = e.keyCode || e.which;
-  const location = e.location;
+
+  const keyCode = event.keyCode || event.which;
+  const location = event.location;
   const mapped = (Object.entries as Entries<Codes, KeyboardKey[]>)(KeyboardMapKeys).find(([, set]) =>
     set.some(([_keyCode, _location]) => _keyCode === keyCode && _location === location),
   );
+
   return mapped ? mapped[0] : Codes.Unidentified;
 }