+++ packages/react-ui/test-setup.js
@@ -1,11 +1,21 @@
+/* eslint-disable react/display-name */
 /* eslint-disable max-len,react/no-deprecated */
 import 'core-js/stable';
 import '@testing-library/jest-dom';
+import { configure } from '@testing-library/dom';
 import React from 'react';
 import Enzyme from 'enzyme';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 
+configure({
+  testIdAttribute: 'data-tid',
+});
+
 Enzyme.configure({ adapter: new Adapter() });
+
+configure({
+  testIdAttribute: 'data-tid',
+});
 
 jest.mock('react-focus-lock', () => (props) => <div>{props.children}</div>);
 jest.mock('lodash.debounce', () =>
@@ -14,6 +24,13 @@ jest.mock('lodash.debounce', () =>
     return fn;
   }),
 );
+jest.mock('react-dom', () => {
+  const originalModule = jest.requireActual('react-dom');
+  return {
+    ...originalModule,
+    findDOMNode: jest.fn(originalModule.findDOMNode),
+  };
+});
 window.matchMedia = jest.fn().mockImplementation((query) => {
   return {
     matches: false,